requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.18
Pygments==2.8.1
PyJWT==2.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
router==0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
django-cors-headers
